move data file commands out of the cypress global name space
